modules/location-ingester/requirements.txt
kafka-python==2.0.2
Faker==13.7.0
grpcio-tools==1.56.2